app/requirements.txt
django==2.2.6
psycopg2
django-debug-toolbar
gunicorn
pytest
pytest-django
pytest-cov
setuptools
ipdb
djangorestframework
django-rest-auth
pillow
djangorestframework_simplejwt
celery==4.3.0
redis==3.2.1
django-celery-beat
sentry-sdk==0.11.2
pyspy
django-rest-knox
flake8
dj-database-url
whitenoise
django-heroku
coveralls
cloudinary==1.18.2